The command predict renamed to analyse
