+++ cmd/root.go
@@ -38,7 +38,7 @@ func init() {
 	app.Init()
 
 	cmdFactory := cli.NewCommandsRunner(app, os.Stdin, os.Stdout, os.Stderr)
-	rootCmd.AddCommand(cmdFactory.NewCmdPredict())
+	rootCmd.AddCommand(cmdFactory.NewCmdAnalyse())
 	rootCmd.AddCommand(cmdFactory.NewCmdBrain())
 	rootCmd.AddCommand(cmdFactory.NewCmdTrain())
